# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TravelerPreferencesSkipped = {[key: string]: boolean};

export interface TravelerPreferences {
  /**
     * @maxItems 6
     * @items.maxLength 30
     */
  vibes: string[];
  /** @maxLength 30 */
  companions: string;
  /**
     * @maxItems 5
     * @items.maxLength 30
     */
  interests: string[];
  /** @maxLength 30 */
  budget: string;
  /**
     * @maxItems 3
     * @items.maxLength 30
     */
  diet: string[];
  skipped: TravelerPreferencesSkipped;
}

export interface TravelerProfile {
  /** @nullable */
  user_id: string | null;
  preferences: TravelerPreferences | null;
}

export interface DailyCost {
  /** @minimum 0 */
  budget: number;
  /** @minimum 0 */
  mid: number;
  /** @minimum 0 */
  lux: number;
}

export interface CountryCityCard {
  slug: string;
  name: string;
  /** @maxLength 100 */
  identity_line: string;
  /**
     * @maxItems 5
     * @items.maxLength 30
     */
  best_for: string[];
  /**
     * @minimum 1
     * @maximum 30
     */
  suggested_days_min: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  suggested_days_max: number;
}

export interface SuggestedRoute {
  /** @maxLength 80 */
  name: string;
  /**
     * @minItems 1
     * @maxItems 8
     */
  cities: string[];
  /**
     * @minimum 1
     * @maximum 30
     */
  days: number;
}

export interface CountryDish {
  /** @maxLength 80 */
  name: string;
  /** @maxLength 140 */
  note: string;
  /** @maxLength 80 */
  where?: string;
}

export type CountryResponseVisaStatusIn = typeof CountryResponseVisaStatusIn[keyof typeof CountryResponseVisaStatusIn];


export const CountryResponseVisaStatusIn = {
  visa_free: 'visa_free',
  voa: 'voa',
  e_visa: 'e_visa',
  embassy: 'embassy',
} as const;

export type CountryResponseMonthlyItem = {
  /**
     * @minimum 1
     * @maximum 12
     */
  month: number;
  label: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  score: number;
};

export interface CountryResponse {
  slug: string;
  name: string;
  region: string;
  /** @maxLength 420 */
  known_for_summary: string;
  visa_status_in: CountryResponseVisaStatusIn;
  /** @minimum 0 */
  visa_cost_inr: number;
  /** @maxLength 300 */
  visa_process_notes: string;
  currency: string;
  daily_cost_inr: DailyCost;
  /**
     * @minItems 1
     * @maxItems 12
     * @items.minimum 1
     * @items.maximum 12
     */
  best_months: number[];
  /** @maxLength 240 */
  domestic_transport: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  safety_score: number;
  /**
     * @minItems 2
     * @maxItems 4
     */
  suggested_routes: SuggestedRoute[];
  /**
     * @minItems 1
     * @maxItems 20
     */
  cities: CountryCityCard[];
  /**
     * @minItems 4
     * @maxItems 6
     */
  top_dishes: CountryDish[];
  /**
     * @minItems 12
     * @maxItems 12
     */
  monthly: CountryResponseMonthlyItem[];
  last_verified: string;
  generated_at: string;
}

export interface GenerateCountryRequest {
  /** @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$ */
  countrySlug: string;
  force?: boolean;
}

export type CityAreaCostBand = typeof CityAreaCostBand[keyof typeof CityAreaCostBand];


export const CityAreaCostBand = {
  budget: 'budget',
  mid: 'mid',
  lux: 'lux',
} as const;

export interface CityArea {
  /** @maxLength 60 */
  name: string;
  /** @maxLength 160 */
  character: string;
  /** @maxItems 5 */
  good_for: string[];
  cost_band: CityAreaCostBand;
}

export type CityPlaceType = typeof CityPlaceType[keyof typeof CityPlaceType];


export const CityPlaceType = {
  activity: 'activity',
  restaurant: 'restaurant',
  stay: 'stay',
  day_trip: 'day_trip',
} as const;

export type CityPlaceCostBandInr = typeof CityPlaceCostBandInr[keyof typeof CityPlaceCostBandInr];


export const CityPlaceCostBandInr = {
  budget: 'budget',
  mid: 'mid',
  lux: 'lux',
} as const;

export type CityPlaceSuitsItem = typeof CityPlaceSuitsItem[keyof typeof CityPlaceSuitsItem];


export const CityPlaceSuitsItem = {
  solo: 'solo',
  partner: 'partner',
  friends: 'friends',
  family: 'family',
  parents: 'parents',
} as const;

export type CityPlaceIndoorOutdoor = typeof CityPlaceIndoorOutdoor[keyof typeof CityPlaceIndoorOutdoor];


export const CityPlaceIndoorOutdoor = {
  indoor: 'indoor',
  outdoor: 'outdoor',
  both: 'both',
} as const;

export interface CityPlace {
  id: string;
  /** @minLength 1 */
  google_place_id: string;
  type: CityPlaceType;
  name: string;
  /** @maxLength 120 */
  description: string;
  address: string;
  cost_band_inr: CityPlaceCostBandInr;
  /**
     * @minimum 0
     * @nullable
     */
  price_inr?: number | null;
  /** @minimum 15 */
  avg_duration_mins: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  physical_intensity: number;
  best_time_of_day: string;
  suits: CityPlaceSuitsItem[];
  veg_friendly: boolean;
  indoor_outdoor: CityPlaceIndoorOutdoor;
  vibe_tags: string[];
  is_signature: boolean;
  /** @nullable */
  booking_url?: string | null;
  /** @nullable */
  affiliate_partner?: string | null;
  bookable: boolean;
  /** @nullable */
  travel_time_mins?: number | null;
  /** @nullable */
  transport_mode?: string | null;
  /** @nullable */
  overnight_recommended?: boolean | null;
  last_verified: string;
}

export interface CityDish {
  name: string;
  /** @maxLength 90 */
  note: string;
  veg: boolean;
}

export type CityTipType = typeof CityTipType[keyof typeof CityTipType];


export const CityTipType = {
  transport: 'transport',
  payment: 'payment',
  connectivity: 'connectivity',
  etiquette: 'etiquette',
  safety: 'safety',
  timing: 'timing',
  booking: 'booking',
} as const;

export interface CityTip {
  type: CityTipType;
  /** @maxLength 70 */
  title: string;
  /** @maxLength 180 */
  body: string;
}

export interface TravellerNote {
  attribution: string;
  /** @maxLength 240 */
  body: string;
  created_at: string;
}

export interface GettingAroundItem {
  mode: string;
  /** @maxLength 160 */
  verdict: string;
}

export interface CityResponse {
  country_slug: string;
  country_name: string;
  slug: string;
  name: string;
  /** @maxLength 100 */
  identity_line: string;
  /** @maxLength 420 */
  known_for_summary: string;
  /** @minimum 1 */
  ideal_days_min: number;
  /** @minimum 1 */
  ideal_days_max: number;
  daily_cost_inr: DailyCost;
  /**
     * @maxItems 12
     * @items.minimum 1
     * @items.maximum 12
     */
  best_months: number[];
  /**
     * @minimum 1
     * @maximum 5
     */
  safety_score: number;
  getting_in: string;
  /**
     * @minItems 4
     * @maxItems 5
     */
  getting_around: GettingAroundItem[];
  /**
     * @minItems 3
     * @maxItems 5
     */
  areas: CityArea[];
  /**
     * @minItems 5
     * @maxItems 5
     */
  signature_places: CityPlace[];
  /** @maxItems 20 */
  activities: CityPlace[];
  /** @maxItems 3 */
  dishes: CityDish[];
  /** @maxItems 8 */
  restaurants: CityPlace[];
  /** @maxItems 4 */
  day_trips: CityPlace[];
  /** @maxItems 8 */
  tips: CityTip[];
  /** @maxItems 6 */
  stays: CityPlace[];
  /** @maxItems 5 */
  traveller_notes: TravellerNote[];
  last_verified: string;
  generated_at: string;
  cached: boolean;
  /** @minimum 0 */
  dropped_count: number;
}

export interface CityGenerationInput {
  /** @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$ */
  citySlug: string;
  /** @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$ */
  countrySlug: string;
  force?: boolean;
}

export interface CityGenerationResult {
  cityId: string;
  cached: boolean;
  /** @minimum 0 */
  droppedCount: number;
}

export type ApiErrorError = {
  code: string;
  message: string;
};

export interface ApiError {
  error: ApiErrorError;
}