# lib/api-zod/src/generated/types/cityArea.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CityAreaCostBand } from './cityAreaCostBand';

export interface CityArea {
  /** @maxLength 60 */
  name: string;
  /** @maxLength 160 */
  character: string;
  /** @maxItems 5 */
  good_for: string[];
  cost_band: CityAreaCostBand;
}