# lib/api-zod/src/generated/types/cityDish.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CityDish {
  name: string;
  /** @maxLength 90 */
  note: string;
  veg: boolean;
}