# lib/api-zod/src/generated/types/cityGenerationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CityGenerationResult {
  cityId: string;
  cached: boolean;
  /** @minimum 0 */
  droppedCount: number;
}