# lib/api-zod/src/generated/types/countryCityCard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CountryCityCard {
  slug: string;
  name: string;
  /** @maxLength 100 */
  identity_line: string;
  /**
     * @maxItems 5
     * @items.maxLength 30
     */
  best_for: string[];
  /**
     * @minimum 1
     * @maximum 30
     */
  suggested_days_min: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  suggested_days_max: number;
}