# lib/api-zod/src/generated/types/suggestedRoute.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SuggestedRoute {
  /** @maxLength 80 */
  name: string;
  /**
     * @minItems 1
     * @maxItems 8
     */
  cities: string[];
  /**
     * @minimum 1
     * @maximum 30
     */
  days: number;
}