# lib/api-zod/src/generated/types/travellerNote.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TravellerNote {
  attribution: string;
  /** @maxLength 240 */
  body: string;
  created_at: Date;
}